qcommon: C++ compatibility for cm_patch

+++ code/qcommon/cm_patch.c
@@ -497,10 +497,10 @@ static qboolean CM_PlaneEqual( const patchPlane_t *p, const float plane[4], int 
 	float invplane[4];
 
 	if (
-	   fabs(p->plane[0] - plane[0]) < NORMAL_EPSILON
-	&& fabs(p->plane[1] - plane[1]) < NORMAL_EPSILON
-	&& fabs(p->plane[2] - plane[2]) < NORMAL_EPSILON
-	&& fabs(p->plane[3] - plane[3]) < DIST_EPSILON )
+	   fabs((double)(p->plane[0] - plane[0])) < NORMAL_EPSILON
+	&& fabs((double)(p->plane[1] - plane[1])) < NORMAL_EPSILON
+	&& fabs((double)(p->plane[2] - plane[2])) < NORMAL_EPSILON
+	&& fabs((double)(p->plane[3] - plane[3])) < DIST_EPSILON )
 	{
 		*flipped = qfalse;
 		return qtrue;
@@ -510,10 +510,10 @@ static qboolean CM_PlaneEqual( const patchPlane_t *p, const float plane[4], int 
 	invplane[3] = -plane[3];
 
 	if (
-	   fabs(p->plane[0] - invplane[0]) < NORMAL_EPSILON
-	&& fabs(p->plane[1] - invplane[1]) < NORMAL_EPSILON
-	&& fabs(p->plane[2] - invplane[2]) < NORMAL_EPSILON
-	&& fabs(p->plane[3] - invplane[3]) < DIST_EPSILON )
+	   fabs((double)(p->plane[0] - invplane[0])) < NORMAL_EPSILON
+	&& fabs((double)(p->plane[1] - invplane[1])) < NORMAL_EPSILON
+	&& fabs((double)(p->plane[2] - invplane[2])) < NORMAL_EPSILON
+	&& fabs((double)(p->plane[3] - invplane[3])) < DIST_EPSILON )
 	{
 		*flipped = qtrue;
 		return qtrue;
@@ -533,13 +533,13 @@ static void CM_SnapVector( vec3_t normal ) {
 
 	for (i=0 ; i<3 ; i++)
 	{
-		if ( fabs(normal[i] - 1) < NORMAL_EPSILON )
+		if ( fabs((double)(normal[i] - 1)) < NORMAL_EPSILON )
 		{
 			VectorClear (normal);
 			normal[i] = 1;
 			break;
 		}
-		if ( fabs(normal[i] - -1) < NORMAL_EPSILON )
+		if ( fabs((double)(normal[i] - -1)) < NORMAL_EPSILON )
 		{
 			VectorClear (normal);
 			normal[i] = -1;
@@ -1644,7 +1644,7 @@ void CM_TraceThroughPatchCollide( traceWork_t *tw, const struct patchCollide_s *
 			else {
 				// NOTE: this works even though the plane might be flipped because the bbox is centered
 				offset = DotProduct( tw->offsets[ pp->signbits ], plane );
-				plane[3] += fabs(offset);
+				plane[3] += fabs((double)(offset));
 				VectorCopy( tw->start, startp );
 				VectorCopy( tw->end, endp );
 			}
@@ -1764,7 +1764,7 @@ qboolean CM_PositionTestInPatchCollide( traceWork_t *tw, const struct patchColli
 			else {
 				// NOTE: this works even though the plane might be flipped because the bbox is centered
 				offset = DotProduct( tw->offsets[ pp->signbits ], plane);
-				plane[3] += fabs(offset);
+				plane[3] += fabs((double)(offset));
 				VectorCopy( tw->start, startp );
 			}
 
@@ -1870,7 +1870,7 @@ void CM_DrawDebugSurface( void (*drawPoly)(int color, int numPoints, float *poin
 				else v1[n] = mins[n];
 			} //end for
 			VectorNegate(plane, v2);
-			plane[3] += fabs(DotProduct(v1, v2));
+			plane[3] += fabs((double)(DotProduct(v1, v2)));
 			//*/
 
 			w = BaseWindingForPlane( plane,  plane[3] );
@@ -1902,7 +1902,7 @@ void CM_DrawDebugSurface( void (*drawPoly)(int color, int numPoints, float *poin
 					else v1[n] = mins[n];
 				} //end for
 				VectorNegate(plane, v2);
-				plane[3] -= fabs(DotProduct(v1, v2));
+				plane[3] -= fabs((double)(DotProduct(v1, v2)));
 
 				ChopWindingInPlace( &w, plane, plane[3], 0.1f );
 			}
